RAG_Documents_Q&A_with_Groq/requirements.txt
langchain
langchain-groq
streamlit
langchain-huggingface
langchain-core
langchain_community
python-dotenv
sentence_transformers
torch==2.2.2